lte-gateway: fix version and incomplete cmds
This changes the use of hard-coded FW_VERSION_* macros
with actual NCS values. It moves printing of this to
one place.

It also ensures bare 'at', 'fota', and 'ble fota'
commands do not crash.

It also ensures the shell prompt is not changed
if a single 'at' command is given.

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>