Rewrite to replace ATv1 base with Multi Service Sample
This commit replaces the previous version with a new version
built on top of the nRF Connect SDK v.2.9.99 on 2025-02-27.

It fixes compatibility issues with v.29.9.99, and replaces all
the base code from the very old Asset Tracker v1 with the better
and newer nRF Cloud Multi Service Sample (MSS).

Due to time constraints this new gateway is not completely
cleaned up to remove unnecessary code from MSS.

And, due to the potentially coming removal of BLE support from
nRF Cloud, this will require modifications in the future to
work as a single normal device using the standard d2c and c2d
topics, rather than the g2c and c2g gateway topics. Further,
this future version could simplify the command and control
protocol defined here, which was appropriate for a JavaScript-
based mobile app, but not very efficient otherwise:
https://github.com/nRFCloud/gateway-common/wiki/Example-Messages

Since nRF Cloud supports custom topic names, those could be
used in the future for each BLE device connected to the gateway:
https://docs.nordicsemi.com/bundle/nrf-cloud/page/APIs/MQTT/Topics.html#additional-topics-supported-by-nrf-cloud
where the topic would include the device's MAC address.

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>